Remove unnecessary type declarations in iter example

+++ examples/iter.rs
@@ -2,7 +2,7 @@ extern crate froggy;
 
 fn main() {
     let mut storage: froggy::Storage<i32> =
-        [5 as i32, 7, 4, 6, 7].iter().cloned().collect();
+        [5, 7, 4, 6, 7].iter().cloned().collect();
     println!("Initial array:");
     for value in storage.iter() {
         println!("Value {}", *value);
@@ -11,7 +11,7 @@ fn main() {
         let item = storage.iter().find(|v| **v == 4).unwrap();
         storage.pin(&item)
     };
-    storage[&ptr] = 350 as i32;
+    storage[&ptr] = 350;
     println!("Array after change:");
     for value in &storage {
         println!("Value {}", *value);
